refactor(schemas): migrate Phase 22 endpoint schemas — 6 files, 18 classes (#6042)
- diagnostics.py (3) → schemas_system.py
- graph_rag.py (3) → schemas_knowledge.py (v1→v2 validator)
- cache_management.py (3) → schemas_system.py
- feature_flags.py (3) → schemas_system.py (imports EnforcementMode)
- documents.py (3) → schemas_knowledge.py
- embeddings.py (3) → schemas_knowledge.py
